Improvide Dynamic plugin structure and generation
Modify Cmake to automatically generate plugins
Move respective plugins in specific directories
